requirements.txt
click>=8.1
anthropic>=0.40
httpx>=0.27
trafilatura>=1.12
pydantic>=2.5
pydantic-settings>=2.0
jinja2>=3.1
openpyxl>=3.1
pandas>=2.0
python-dotenv>=1.0
rich>=13.7
tenacity>=8.2
ddgs>=9.0
openai>=1.17.0
fastapi>=0.115
uvicorn[standard]>=0.32
python-multipart>=0.0.18
sse-starlette>=2.1
aiosqlite>=0.20
